+++ changelog.md
@@ -1,5 +1,6 @@
 # Version 2024.10.9 (2024-10-16)
 
+- Add reportValueUsage and getLinks to client
 - Add version to code
 
 # Version 2024.10.8 (2024-10-15)